requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
curl_cffi==0.13.0
feedparser==6.0.12
frozendict==2.4.7
idna==3.11
multitasking==0.0.12
narwhals==2.18.1
numpy==2.0.2
packaging==26.0
pandas==2.3.3
peewee==4.0.2
platformdirs==4.4.0
plotly==6.6.0
protobuf==6.33.6
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
yfinance==1.2.0
Flask==3.1.0
